+++ NEWS.md
@@ -5,6 +5,8 @@
 - Addition of `groupMsFragments()` function. This function generates a 
   `numeric` index grouping MSn (MS level > 1) spectra with their corresponding
   MS1 spectra based on acquisition order.
+- Fix on `cbind2()` so that the output is a `Spectra` object and not the 
+  backend. 
 
 ## Change 1.19.4